pyproject.toml
[tool.poetry]
name = "jotbox"
version = "0.1.0"
description = ""
authors = ["Steinthor Palsson <steini90@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.5.1"
pyjwt = "^1.7.1"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
flake8 = "^3.8.3"
black = "^19.10b0"
aredis = "^1.1.8"
databases = "^0.3.2"
sqlalchemy-stubs = "^0.3"
aiopg = "^1.0.0"
mkdocs = "^1.1.2"
mkdocs-material = "^5.4.0"
markdown-include = "^0.5.1"
pygments = "^2.6.1"
pymdown-extensions = "^7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
